+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Support for the `WEAK_NATIVEINT` symbol, which is defined from Delphi 12 onward.
+
 ### Changed
 
 - Exclude types annotated with attributes in `UnusedType`.